fix: OPENAI_API_KEY is taken from the environment instead of .env
This is arguably a matter of taste, but in my case, the `.env` values should
override the values set in `~/.profile`.

This disables the ability to set the environment on the command-line, but to
those who would take an issue with that, please change the `.env` temporarily
instead.

For most users this will not matter one way or another.